requirements.txt
paho-mqtt
pyyaml
schema
